# docs/review-api-openapi.yaml
openapi: 3.0.3
info:
  title: Review API
  description: API untuk mengelola ulasan produk dengan fitur CRUD lengkap, statistik, dan manajemen admin
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://your-domain.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /admin/reviews:
    get:
      summary: Get All Reviews (Admin)
      description: Mengambil semua ulasan dengan informasi lengkap untuk admin
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Halaman
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Jumlah item per halaman
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          description: Field untuk sorting
          required: false
          schema:
            type: string
            enum: [created_at, rating, updated_at]
            default: created_at
        - name: sortOrder
          in: query
          description: Urutan sorting
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: verified
          in: query
          description: Filter berdasarkan status verifikasi
          required: false
          schema:
            type: boolean
        - name: rating
          in: query
          description: Filter berdasarkan rating
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReviewListResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update Review (Admin)
      description: Admin dapat memperbarui ulasan apapun
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateReviewRequest'
      responses:
        '200':
          description: Review updated successfully by admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReviewResponse'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Review:
      type: object
      properties:
        id:
          type: integer
          example: 1
        productId:
          type: integer
          example: 1
        userId:
          type: string
          format: uuid
          example: "a4bc7b55-bee9-4f13-8486-9cb8bb92be29"
        userName:
          type: string
          example: "John Doe"
        userAvatar:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Produk sangat bagus!"
        date:
          type: string
          format: date
          example: "2024-01-15"
        verified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    AdminReview:
      allOf:
        - $ref: '#/components/schemas/Review'
        - type: object
          properties:
            productName:
              type: string
              example: "Laptop Gaming ASUS"
            userEmail:
              type: string
              format: email
              example: "john@example.com"
...
    AdminUpdateReviewRequest:
      type: object
      required:
        - reviewId
      properties:
        reviewId:
          type: integer
          example: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: string
          minLength: 10
          maxLength: 500
          example: "Review diperbarui oleh admin"
        verified:
          type: boolean
          example: true
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
    AdminReviewListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/AdminReview'
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    AdminReviewResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Review berhasil diperbarui oleh admin"
        data:
          $ref: '#/components/schemas/AdminReview'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Rating harus antara 1-5"
...